// Frontend/client-interop/interop-query-response.d.ts
/**
 * This is a TypeGen auto-generated file.
 * Any changes made to this file can be lost when this file is regenerated.
 */
import { InteropQuery } from "./interop-query";
import { ResponseStatus } from "./response-status";
export declare class InteropQueryResponse {
    ticks: number;
    query: InteropQuery;
    status: ResponseStatus;
    result: Object;
    request: Object;
    message: string;
}